m113: Update gpu/
